Fix a "conflicting types for 't9_file_size'" build error
```
../src/t9/io.c:55:1: error: conflicting types for 't9_file_size'
t9_file_size(const char *const path, size_t *const size) {
^
../include/t9/io.h:63:1: note: previous declaration is here
t9_file_size(const char *const path, uint64_t *const size);
^
```

+++ include/t9/io.h
@@ -60,7 +60,7 @@ t9_file_exists(const char *const path, uint8_t *const exists);
  * @return T9_SUCCESS on success, T9_FAILURE otherwise.
  */
 t9_error_t
-t9_file_size(const char *const path, uint64_t *const size);
+t9_file_size(const char *const path, size_t *const size);
 
 /*!
  * Read the content of a file to memory.
